+++ cogeo_mosaic/backends/web.py
@@ -13,7 +13,7 @@ import httpx
 from cachetools import TTLCache, cached
 from cachetools.keys import hashkey
 
-from cogeo_mosaic.backends.base import BaseBackend
+from cogeo_mosaic.backends.base import MosaicJSONBackend
 from cogeo_mosaic.backends.utils import _decompress_gz
 from cogeo_mosaic.cache import cache_config
 from cogeo_mosaic.errors import _HTTP_EXCEPTIONS, MosaicError
@@ -21,7 +21,7 @@ from cogeo_mosaic.mosaic import MosaicJSON
 
 
 @attr.s
-class HttpBackend(BaseBackend):
+class HttpBackend(MosaicJSONBackend):
     """Http/Https Backend Adapter"""
 
     # Because the HttpBackend is a Read-Only backend, there is no need for